refactor: remove old Ant Design Button and migrate to UI library Button
Remove the deprecated Ant Design Button component at
components/Button/Button/Button.tsx and migrate all 30 consumers to use
the existing UI library Button from components/Button.

Prop mapping applied:
- type="primary" → removed (already default kind="primary" emphasis="high")
- type="default" → kind="secondary" emphasis="high"
- type="text" → kind="secondary" emphasis="low"
- danger + type="primary" → kind="danger"
- iconButton → removed (not applicable to new Button)
- trackProperties → trackingProperties

Fixes #5753
