+++ frontend/src/components/ErrorState/RequestAccess/RequestAccess.tsx
@@ -1,4 +1,4 @@
-import Button from '@components/Button/Button/Button'
+import { Button } from '@components/Button'
 import { toast } from '@components/Toaster'
 import { useRequestAccessMutation } from '@graph/hooks'
 import { useParams } from '@util/react-router/useParams'
@@ -15,7 +15,6 @@ const RequestAccess = () => {
 			small
 			trackingId="ErrorStateRequestAccess"
 			disabled={sentAccessRequest}
-			type="primary"
 			onClick={async () => {
 				try {
 					await requestAccess({
