+++ frontend/src/pages/WorkspaceSettings/FieldsForm/FieldsForm.tsx
@@ -7,7 +7,7 @@ import { useParams } from '@util/react-router/useParams'
 import React, { useState } from 'react'
 
 import commonStyles from '../../../Common.module.css'
-import Button from '../../../components/Button/Button/Button'
+import { Button } from '@components/Button'
 import { CircularSpinner } from '../../../components/Loading/Loading'
 import styles from './FieldsForm.module.css'
 
@@ -105,7 +105,6 @@ export const FieldsForm: React.FC<Props> = ({
 								isWorkspace ? 'Workspace' : 'Project'
 							}Update`}
 							htmlType="submit"
-							type="primary"
 							className={commonStyles.submitButton}
 							disabled={formDisabled}
 						>
